Rename custom detail to extended detail.
Custom fields inside details will be disabled soon, as it can't properly handle the case when the schema of the back-end has been changed. All the data which are not pre-defined in the detail leaf classes should be exposed through QOrganizerItemExtendedDetail.

NOTE: The Customized enumerator in QDeclarativeOrganizerItemDetail will be changed later.

Change-Id: Idadb914b2a3bc65c91055ac20a04264b08e349d1
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Xizhi Zhu <xizhi.zhu@nokia.com>