Fix issue with details and keys
Part of the details are unique for an item. When such detail is saved to
an item, it updates the one stored previously. Updation was done by
setting the value of the previous one and this lost the sync of detail's
keys in some cases.

1. Instead of setting new value to detail, the whole detail is now
replaced. This keeps the keys in sync.
2. Previously Type-detail deletion was not allowed by this behaviour.
Now there is separate check for this.
3. Updated test cases to go along with the change.

Change-Id: I585e7469fef5dff456bd44b44696bc198c06bc87
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Xizhi Zhu <xizhi.zhu@nokia.com>