Convert casting contact filters to do copy construction
Filters have a base class but they are copied by value, and the copy
constructor in the base class sets the private part to point to the
private part of the filter from the derived class. Therefore, such a
filter object has a type QContactFilter (not the derived class) and
should not be casted down to the derived class.

Change-Id: I750f1d469f782e5b3b53e7ec0869de8ffbe087b0
Reviewed-by: Xizhi Zhu <xizhi.zhu@nokia.com>
Reviewed-by: Cristiano di Flora <cristiano.di-flora@nokia.com>