Remove QContactThumbnail detail.
This change removes the QContactThumbnail detail from contacts API.
From now on, the handling of PHOTO property of a vcard in Versit will
by default cover only the case of a URL value in the PHOTO field.
Base64 encoded PHOTO properties are not added by default anymore to
QContact objects. In case a specific platform requires storing Base64,
this will have to be achieved by implementing platform-specific property
handler plugins, e.g. taking care of storing the encoded thumbnails to
the platform specific media storage location and converting the image to
one or more formats that best fit the target platform.

Change-Id: I46028b6229b04be0da853a5c02a600ae0d7b63ec
Reviewed-by: Kranthi Kuntala <kranthi.kumar-kuntala@nokia.com>