Fixing the handling of "date-only" parameters in Organizer QML API
Setters and getters of some Organizer item detail fields, such as
originalDate in Parent detail, have "date-only" parameters. The type
of the parameter is JavaScript Date, but since only date part of the
Date object was needed, the type of the C++ parameter was QDate.
This resulted in implicit conversions from QDateTime to QDate where
timezone offsets were not handled correctly. Now type of C++
parameters has been changed to QDateTime in order to allow more
control over the conversion between QDateTime and QDate.

Change-Id: I9357b0935bf6a041829bcaaa273dcf9bb662d72e
Reviewed-by: Iiro Kause <iiro.kause@nokia.com>