Enable faster appending of details to QContact objects.
This change introduces a new method into QContact class,
namely QContact::appendDetail(). The method is similar to
saveDetail. The main difference is that the method does not
try to loop over all the already stored details to try
updating an existing detail, but rather just appends the
given detail to the list of stored details.
The goal is to speed up the use cases where a brand new
QContact is created by clients, and then extended by adding
several details to it in a row.
This can be useful e.g. for backend implementors when fetching
data from the back-end specific data storage.
The jsondb backend is also modified by this change so as to take
this new method into usage. The change brings a 6x-8x performance
improvement to the basic json to qcontact conversion logic
in the qcontactjsondbconverter.

Change-Id: I9e7a18e3d56c656d10f3315bd03820632ce87b52
Reviewed-by: Claudio Brunelli <claudio.brunelli@nokia.com>
Reviewed-by: Tommi Anttila <tommi.4.anttila@nokia.com>
Reviewed-by: Kranthi Kuntala <kranthi.kumar-kuntala@nokia.com>