Remove QContactJsonDbEngineData class.
The shared data pointer in jsondb engine class was not actually needed.
This change uses a standard QContactJsonDbRequestHandler pointer inside
the QContactJsonDbEngine class. With this change, a severe memory leak
due to bad usage of that shared pointer (resulting in handler object
not actually deleted) is removed.
This change also disables the copy of QContactJsondDbEngine objects by
removing the implementation of the assignment operator.

Change-Id: I25195f1c254c6a60f6ff6c1686cf364b40d1b3b7
Reviewed-by: Cristiano di Flora <cristiano.di-flora@nokia.com>
Reviewed-by: Mikko Suonio <mikko.suonio@nokia.com>