Make jsondb organizer tests more defensive.
Some of the auto tests that are using the jsondb backend cannot work
without proper partitions.json file. This change makes the tests more
defensive with respect to situations when that file is not deployed
correctly. Instead of just showing a warning (previous case) we let now
the initTestCase fail in case the partitions.json file cannot be
successfully located.

Change-Id: I731ecc94199c98a81daaf9bc060433fd43aa3ac5
Reviewed-by: Pekka Kauppila <pekka.kauppila@nokia.com>
Reviewed-by: Andy Cox <andrew.cox@nokia.com>
Reviewed-by: Päivi Rajala <paivi.rajala@nokia.com>