Fix test failures due to QVariant debug output change
The QVariant debug output for invalid QVariants has changed in Qt.
Before it was "QVariant(void, QVariant::Invalid)", now it's
"QVariant(Invalid)".

This is because void is now a distinct type (QMetaType::Void), and
QVariant::Invalid and QMetaType::Void no longer happen to have the
same integral value.

Adapt the qorganizeritemfilter test accordingly.

Change-Id: I33673f2b94aed1b20a6635fd51dd064b5a3fb0ef
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@nokia.com>