Do not restart the timer after every notification from jsondb
If the timer is started on every notification, a notification arriving
before the previous timer has lapsed causes the timer to be restarted
and the signal delivery is postponed by another timeout duration.
In practice, this made performance tests creating items in a loop
to collect all notifications in one large batch and only a single
added/changed/removed notification was delivered.

Change-Id: I539445174acfab73942688cd4ad5527ce58dd260
Reviewed-by: pengwu <peng.1.wu@nokia.com>
Reviewed-by: Tommi Anttila <tommi.4.anttila@nokia.com>
Reviewed-by: Cristiano di Flora <cristiano.di-flora@nokia.com>