Don't try to access item details list if it's empty.
This change adds an extra check to setType implementation so as to let
it tolerate the case where the copy-constructor or assignment of
an item is called for an item having no details in its internal
details list. This condition has been randomly seen in flacky crashes
of qorganizermanager test runs upon random failures of jsondb server.

Change-Id: Iaa01b06a315e6256b9ca56c18ce29a11a2042f05
Reviewed-by: Päivi Rajala <paivi.rajala@nokia.com>