Refactoring Organizer detailFilter class.
This patch changes the interface and behaviour of
QOrganizerItemDetailFilter class (and of the corresponding
QML DetailFilter element).
The previous class and element, based on single
field / value pair matching  are now renamed as
QOrganizerItemDetailFieldFilter and DetailFieldFilter
respectively.
The new detail filter implementation provides a
convenince filter for exact detail-to-detail matching.
No MatchFlags or wildcards are supported. Matching items
must have a detail that is identical to that of
the filter.
The old interface of DetailFilter is still available in the
refactored class / element, but it is deprecated and will be
removed in a separate change.

Change-Id: I3911f76476fecf10a4813509dc2cbbda3057aa56
Reviewed-by: Cristiano di Flora <cristiano.di-flora@nokia.com>