Finetune Organizer storage location errors
StorageLocationsNotExistingError is removed from
QOrganizerManager. It is replaced by 2 new error codes:

InvalidStorageLocationError - This is returned in cases
when non-mandatory but not existing storage location is
tried to be accessed.

MissingPlatformRequirementsError - This is for cases
when organizer cannot operate due missing organizer backend
requirements. For jsondb backend such case is that User-
partition is not present at all.

Change-Id: I8f1aa51dcd21a3a3b3f8e939ba22f18f2180869f
Reviewed-by: Päivi Rajala <paivi.rajala@nokia.com>