Adjust handling of fetch request ids in the organizer model
The backend may complete the request and call
onFetchItemsRequestStateChanged() within the call to start(), so
the fetch request id needs to be in the hash before start() is called.

Also, model should remove the request from the hash when the request
is finished. Otherwise, those will stay there and the hash will grow
out of bounds.

Change-Id: Ib0e3003445d98c659700d1406c8a4c540cd4bb9d
Reviewed-by: Iiro Kause <iiro.kause@nokia.com>
Reviewed-by: Päivi Rajala <paivi.rajala@nokia.com>