Fix the qml list interface
The qml list interface was changed, so that there is now a read only
constructor and the default properties are removed. This is because the
original interface was very runtime centric, like other parts in qml.
The new interface was changed for the needs of the qml designer and
debugger.

Task-number: QTBUG-9390

Change-Id: I52e9e946e4986107677a130f4e46e820f000b55e
Reviewed-by: Thomas Hartmann <Thomas.Hartmann@digia.com>