Fix Q_DECLARE_METATYPE uses.
Remove one use of it which is unnecessary (The QList of an already
declared metatype), and move the other to the top of the file,
particularly above the use of qRegisterMetaType with it.

The use of qRegisterMetaType above the use of Q_DECLARE_METATYPE in
the file caused an ODR violation.

Also use the nullary overload of qRegisterMetaType instead of using
a typo-prone string. Using the non-nullary overload is what caused
the ODR violation in the first place.

Change-Id: I60f5060af85d76489c74ecc08d085941183f9532
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@digia.com>