Work around issues in qtpim. The primary target is to unblock integration
Currently all of qtpim integration is dead as tests fail left and
right. Target is to unblock qtpim/dev branch running against
a CI targeting qtbase/stable. This is the as is state
and promises to be a bit safer due to the stability difference
between stable and dev at this point in time. Later we will change this
to the proper target (being the dev branch).

Once everything is unlocked we will continue regular integration using
qtpim/dev and closing down master.

Work arounds so far:

- Skip qmlorganizer::OrganizerRecurrenceTests::test_changeTimePeriod(row 0)

Change-Id: I0033537cfeab7e9986d8b0f8b0d0a94baf643b63
Reviewed-by: Konstantin Ritt <ritt.ks@gmail.com>
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>