Extend changed signals to optionally report changed types
Add a parameter to the contactsChanged and itemsChanged signals, which
allow a manager to optionally report the extent of changes involved in
a contact or organizer item modification. If no changes are explicitly
reported by the manager, the receiver must assume that any or all
details of the contacts or organizer items may have been modified.

Change-Id: Ie8a0d051c3f82a5911dea68213d78c2b173e762e
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>
Reviewed-by: Christopher Adams <chris.adams@jollamobile.com>