Add QContactType::TypeFacet value
Allow a contacts backend to provide explicit support for composition
of contact representations into a single composite contact.

A single logical contact can be represented in a QtContacts system
as multiple separate elements, each created and maintained by separate
software systems. To support this use-case, provide the 'Facet'
contact type, which has the same structure as a TypeContact contact,
but different semantics.  A facet is defined as the portion of an
individual contact known by any one system, which is probably incomplete.
For example, an email service will be able to know the name and email
address of a contact from message metadata, but has no access to phone
numbers or mailing addresses. By storing contact details as facets, it
explicitly allows them to be aggregated into more comprehensive contact
representations.

Facet contacts are made available to clients as regular contacts by
producing contacts of type TypeContact, which aggregate the details of
the facets together into a composite whole. The correspondence between
facets and their composing contacts is defined by QContactRelationship
records with type 'Aggregates'.

The semantics governing the behavior of facets and composite contacts
when modified or deleted are left to be defined by each backend
that implements the Facet type.

Change-Id: I486e2d9140fca5d306dc56fb5de9bb6ca2327686
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>