[versit] Ensure VCARD output is deterministic
Sort the parameters emitted by the VCARD writers so ensure that the
results are deterministic for testing purposes. The specification
does not constrain ordering, so this change is required only to
make the test code simple and robust.

Also sort some QDebug output values for the same reason.

Task-number: QTBUG-25382

Change-Id: Ifcebc4c13fc1c3b6cb963235c55b83cee1a65571
Reviewed-by: Alex Blasche <alexander.blasche@digia.com>