PIM ID local components do not need to be human-readable
The PIM ID objects compose two elements: the URI of the manager that
owns them, and a local component meaningful only to that manager. The
URI is composed from human-readable elements and benefits from a
string representation, but the local element has no requirement on
readability, and can consume less storage by using a binary
representation.

Change-Id: I56d553f6d9debf0486310688006b0be8728c1c90
Reviewed-by: Konstantin Ritt <ritt.ks@gmail.com>
Reviewed-by: Christopher Adams <chris.adams@jollamobile.com>