Fix crash in the declarative contact model
Make sure we don't try to move contacts to an invalid index when
updating the contacts.
This change protects against moving contacts, in the declarative contact
model, to an invalid index. In most cases this will never happen, but
if the pending contacts list contains duplicates, then  it's possible
that the index, used to move existing contacts, is out of range.
The worst case scenario with this change, is that a contact will be
moved more then once, that is, once for each duplicate.

Change-Id: I567ea5d9a195819550c98b0da9f45aa613cb3a79
Reviewed-by: Renato Araujo Oliveira Filho <renato.filho@canonical.com>
Reviewed-by: Christopher Adams <chris.adams@jollamobile.com>