QtOrganizer: fix date-only value handling to avoid UTC conversion
Some functions take JS Date parameters which can be either intended
to be QDate or QDateTime values.  Previously these functions had always
assumed that the Date parameter was a fully-specified datetime and
performed a toUtc() conversion before extracting the date part.

For example, new Date("January 1, 1986") would be converted to a
QDateTime and then have the local-to-UTC timezone transformation
applied, which could result in an incorrect date.

This commit ensures that in those cases, we detect whether the Date
was most likely constructed as a date-only Date, and in those cases
we immediately extract the date part without performing a toUtc()
conversion.

Change-Id: I8f76cccc2fc27df2115965c78288e16185b1a0c8
Reviewed-by: Matthew Vogt <matthew.vogt@qinetic.com.au>
Reviewed-by: Renato Araujo Oliveira Filho <renato.filho@canonical.com>
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>