QtContacts: detail-leaf-class API improvements
This commit improves the API of various detail leaf classes by adding
support for more commonly-required fields.

It also fixes the QContactGender class so that the enum definitions
are sensible and consistent with the rest of the API.

Finally, it adds support for providing per-detail provenance
information, to enable the Aggregate/Facet contact story.

Change-Id: Idf465f99963d372ec4f0d439cf9f3be8a7823202
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>
Reviewed-by: Matthew Vogt <matthew.vogt@qinetic.com.au>