Fix QtVersit encoding of Gender details
The Gender detail API was tweaked slightly for consistency with the
rest of the API, but the Versit encoder was not updated at the same
time.

This commit ensures that the Versit encoder treats default-constructed
QContactGender details as "Unspecified" gender values, and updates the
unit test to enforce that semantic.

Change-Id: I9f6a75888fbe8497b66ff7ec2104cacd7575fa50
Reviewed-by: Matthew Vogt <matthew.vogt@qinetic.com.au>