Fix breakage caused by QtQml "null" assignment semantics change in 5.8
Previously, assigning a "null" value to a QVariant Q_PROPERTY from
JavaScript within QML would cause the QVariant to contain a value
whose type was QMetaType::VoidStar.  The semantics of this operation
changed in Qt 5.8 so that now the QVariant will contain a value
whose type is QMetaType::Nullptr (a new type added in Qt 5.8).

This commit ensures that we handle this change via #if.

Change-Id: Iaa4238998b0b3f4c4f36f2ff31501eba2ce132d4
Reviewed-by: Jake Petroules <jake.petroules@qt.io>