Add label group field to display label detail
Modern UIs often display contacts in sections grouped by the first
letter of their first (or last, depending on platform and locale
preferences) name.

This commit adds a label group field to the display label detail
which allows a locale-specific label group to be stored for each
contact and exposed to clients.

Change-Id: I40f20fd7e8861e74618c5d0a76fc3c2b8119dfc0
Reviewed-by: Matthew Vogt <matthew.vogt@qinetic.com.au>