Ensure we throw away the BOM in the qversitreader test
The BOM will only be emitted from the very first call to
encoder->fromUnicode(), and thus since we compare the output from
one invocation to the output of another, we need to ensure that
the BOM is not included in the result of either.

However, previously we attempted to do this by passing a null
QString into fromUnicode(), however it seems that now it ignores
such input.  Hence, we need to give it some dummy input in the
first call to fromUnicode() to force it to "spend" its BOM output.

Change-Id: I74d8bd53604044ba1e82d48668d9570505fef491
Reviewed-by: Pekka Vuorela <pvuorela@iki.fi>